rename package 'entities' -> 'dtos'
