+++ src/main/java/VersionCheckWindow.java
@@ -8,8 +8,8 @@ import com.intellij.openapi.wm.ToolWindowFactory;
 import com.intellij.ui.HyperlinkAdapter;
 import com.intellij.ui.content.Content;
 import com.intellij.ui.content.ContentFactory;
-import entities.Dependency;
-import entities.DependencyUpdatesResult;
+import dtos.Dependency;
+import dtos.DependencyUpdatesResult;
 import org.gradle.tooling.GradleConnectionException;
 import org.gradle.tooling.GradleConnector;
 import org.gradle.tooling.ProjectConnection;
